requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2==2.9.9
python-dotenv==1.1.0
# Following package only for development on windows
# psycopg2-binary==2.9.10